RavenDB-24867 - Address PR/spoken comments
